feat(wasm-tools): F4.5.4 port runtime.rs + wrapper.rs + wit to dasclaw_wasm_tools
ADR-152 F4.5.4 Plan A_extended verbatim port (ADR-129 §1.3 红线遵守).

- 物理迁移 runtime.rs / wrapper.rs / wit/tool.wit 到 crates/dasclaw_wasm_tools/
- OAuthTokenResponse / ProxyRefreshTokenRequest / oauth_token_response_from_json
  / refresh_token_via_proxy 上提到 dasclaw_llm_provider::provider::oauth_helpers
- 7 个测试常量 + test_secrets_store() 共享至
  crates/dasclaw_wasm_tools/src/test_credentials.rs
- desktop 旧路径转为薄壳 re-export
- loader.rs + rate_limiter.rs 按 ADR-152 addendum §4.2 暂留 desktop

验证:
- cargo check -p dasclaw_wasm_tools --tests: 0 错 0 警
- cargo check -p dasclaw --tests: 0 错 0 警
- cargo nextest run -p dasclaw_wasm_tools: 160 passed
- cargo clippy --no-deps -p dasclaw_wasm_tools --all-targets -- -D warnings: clean
- check_no_panics: OK
- check_no_new_ironclaw_literal: OK (类 A)

已检查 dasclaw_wasm_tools / dasclaw_llm_provider 是否已有等价实现,结论:Phase 1
(oauth)+ Phase 2 (test fixtures)在前序会话已落,本次 Phase 3 完成搬迁与薄壳。

Closes #742
