+++ crates/dasclaw_wasm_tools/Cargo.toml
@@ -14,18 +14,24 @@ anyhow = "1"
 async-trait = "0.1"
 blake3 = "1"
 chrono = { version = "0.4", features = ["serde"] }
+dirs = "6"
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-native-roots", "stream"] }
+secrecy = "0.10"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "2"
-tokio = { version = "1", features = ["sync"] }
+tokio = { version = "1", features = ["sync", "rt"] }
 tracing = "0.1"
 url = "2"
 urlencoding = "2"
 uuid = { version = "1", features = ["v4", "v5", "serde"] }
 wasmtime = { version = "28", features = ["component-model"] }
+wasmtime-wasi = "28"
 
 # Workspace crates
+dasclaw_llm_provider = { path = "../dasclaw_llm_provider" }
 dasclaw_runtime = { path = "../dasclaw_runtime" }
+dasclaw_safety = { path = "../dasclaw_safety" }
 dasclaw_tool = { path = "../dasclaw_tool" }
 
 # Optional storage backends
@@ -39,5 +45,7 @@ postgres = ["dep:deadpool-postgres", "dep:tokio-postgres", "dasclaw_runtime/post
 libsql = ["dep:libsql", "dasclaw_runtime/libsql"]
 
 [dev-dependencies]
+axum = "0.8"
 secrecy = "0.10"
-tokio = { version = "1", features = ["sync", "rt", "macros"] }
+tempfile = "3"
+tokio = { version = "1", features = ["sync", "rt", "macros", "net"] }
