+++ crates/dasclaw_wasm_tools/src/lib.rs
@@ -1,7 +1,7 @@
 //! WASM sandbox primitives for tool execution.
 //!
 //! This crate hosts the verbatim port of `desktop-client/ironclaw/src/tools/wasm/`
-//! infrastructure modules per [ADR-152 §4.2 addendum] (F4.5.3). The following
+//! infrastructure modules per [ADR-152 §4.2 addendum] (F4.5.3 + F4.5.4). The following
 //! files moved here unchanged (only `use` paths and `pub(crate)` visibility were
 //! adjusted per ADR-129 §1.3.1 fork-able boundary):
 //!
@@ -13,12 +13,15 @@
 //! - `allowlist` — HTTP endpoint allowlist validator
 //! - `storage` — `WasmToolStore` trait + Postgres / libSQL implementations
 //! - `credential_injector` — secret injection into outbound HTTP requests
+//! - `runtime` — `WasmToolRuntime` engine/store lifecycle (F4.5.4)
+//! - `wrapper` — `WasmToolWrapper` Tool impl + WIT bindings (F4.5.4)
+//! - `test_credentials` — shared fixtures consumed by wrapper tests (F4.5.4)
 //!
-//! The modules `wrapper`, `runtime`, `rate_limiter`, and `loader` remain in
-//! `desktop-client/ironclaw/src/tools/wasm/` because they couple to desktop
-//! oauth helpers, the desktop `ToolRegistry`, or shim into desktop modules
-//! outside `wasm/`. See `tools/wasm/mod.rs` for the re-export shim that keeps
-//! `crate::*` working from the desktop side.
+//! The modules `rate_limiter` and `loader` remain in
+//! `desktop-client/ironclaw/src/tools/wasm/` because they couple to the desktop
+//! `ToolRegistry` or shim into desktop modules outside `wasm/`. See
+//! `tools/wasm/mod.rs` for the re-export shim that keeps `crate::*` working from
+//! the desktop side.
 
 pub mod allowlist;
 pub mod capabilities;
@@ -27,16 +30,27 @@ pub mod credential_injector;
 pub mod error;
 pub mod host;
 pub mod limits;
+pub mod runtime;
 pub mod storage;
+pub mod test_credentials;
+pub mod wrapper;
+
+/// WIT interface version exposed to tools (matches `wit/tool.wit`).
+pub const WIT_TOOL_VERSION: &str = "0.3.0";
+
+/// Channel WIT interface version (matches `desktop-client/ironclaw/wit/channel.wit`).
+pub const WIT_CHANNEL_VERSION: &str = "0.3.0";
 
 // Core re-exports (mirror of the historical `tools/wasm/mod.rs` block, minus
-// the four modules that stay in the desktop crate).
+// the two modules that stay in the desktop crate).
 pub use error::WasmError;
 pub use host::{HostState, LogEntry, LogLevel};
 pub use limits::{
     DEFAULT_FUEL_LIMIT, DEFAULT_MEMORY_LIMIT, DEFAULT_TIMEOUT, FuelConfig, ResourceLimits,
     WasmResourceLimiter,
 };
+pub use runtime::{PreparedModule, WasmRuntimeConfig, WasmToolRuntime, enable_compilation_cache};
+pub use wrapper::{OAuthRefreshConfig, WasmToolWrapper};
 
 pub use capabilities::{
     Capabilities, EndpointPattern, HttpCapability, RateLimitConfig, SecretsCapability,
